Update deploy.yml
Updating the deploy file to deploy code changes to the digitalocean droplet CarCatch. This just echos that the deployment was successful on every push to master. This is a proof of concept for connectivity between github and digitalocean.

THIS DOES NOT ACTUALLY DEPLOY THE CODE YET, since there is no code yet...